requirements.txt
asgiref==3.7.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.13
django-cors-headers==4.7.0
django-filter==24.2
django-jalali==7.4.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
Markdown==3.6
mysqlclient==2.2.7
numpy==1.26.4
openpyxl==3.1.2
packaging==23.1
pandas==2.2.2
persiantools==5.3.0
psutil==7.0.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-gettext==5.0
pytz==2024.1
requests==2.32.5
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.5.0
whitenoise==6.5.0
